Revert "wip ctx provider, hook,  methods"
This reverts commit 02601ae2ca39235dd378bb75b12ca5b8476106af.
